fix: master build breaks

+++ vhdbuilder/packer/install-dependencies.sh
@@ -414,12 +414,13 @@ if [[ ${UBUNTU_RELEASE} == "18.04" && ${ENABLE_FIPS,,} == "true" ]]; then
   relinkResolvConf
 fi
 
-# remove snapd, which is not used by container stack
-apt-get purge --auto-remove snapd -y
-
-# update message-of-the-day to start after multi-user.target
-# multi-user.target usually start at the end of the boot sequence
-sed -i 's/After=network-online.target/After=multi-user.target/g' /lib/systemd/system/motd-news.service
+if [[ $OS == $UBUNTU_OS_NAME ]]; then
+  # remove snapd, which is not used by container stack
+  apt-get purge --auto-remove snapd -y
+  # update message-of-the-day to start after multi-user.target
+  # multi-user.target usually start at the end of the boot sequence
+  sed -i 's/After=network-online.target/After=multi-user.target/g' /lib/systemd/system/motd-news.service
+fi
 
 # retag all the mcr for mooncake
 # shellcheck disable=SC2207
@@ -430,11 +431,11 @@ else
   # shellcheck disable=SC2016
   allMCRImages=($(docker images | grep '^mcr.microsoft.com/' | awk '{str = sprintf("%s:%s", $1, $2)} {print str}'))
 fi
-if [[ "${allMCRImages}" == "" ]]; then
+if [[ "${allMCRImages}" == "" && "$OS" == "$UBUNTU_OS_NAME" ]]; then
   echo "we must find some mcr images"
   exit 1
 fi
-for mcrImage in "${allMCRImages[@]}"; do
+for mcrImage in ${allMCRImages[@]+"${allMCRImages[@]}"}; do
   # in mooncake, the mcr endpoint is: mcr.azk8s.cn
   # shellcheck disable=SC2001
   retagMCRImage=$(echo ${mcrImage} | sed -e 's/^mcr.microsoft.com/mcr.azk8s.cn/g')
